third_party/blink/web_tests/platform/mac-retina/external/wpt/webrtc/RTCPeerConnection-perfect-negotiation.https-expected.txt

This is a testharness.js-based test.
PASS Perfect negotiation setup connects
PASS Perfect negotiation setup connects with roles reversed
FAIL Perfect negotiation glare assert_equals: negotiationneeded always fires in stable state expected "stable" but got "have-remote-offer"
FAIL Perfect negotiation glare with roles reversed assert_unreached: Error in iframe: Error: negotiationneeded always fires in stable state expected stable but got have-remote-offer Reached unreachable code
FAIL Perfect negotiation stress glare assert_unreached: Error in iframe: Error: onmessage SLD worked expected answer but got offer Reached unreachable code
FAIL Perfect negotiation stress glare with roles reversed assert_unreached: Error in iframe: Error: negotiationneeded always fires in stable state expected stable but got have-remote-offer Reached unreachable code
FAIL Perfect negotiation stress glare linear assert_unreached: Error in iframe: Error: onmessage SLD worked expected answer but got offer Reached unreachable code
FAIL Perfect negotiation stress glare linear with roles reversed assert_unreached: Error in iframe: Error: negotiationneeded always fires in stable state expected stable but got have-remote-offer Reached unreachable code
Harness: the test ran to completion.
